jeca: introduce customfield handlers
You can now have a directory with python code to interpret the custom
fields and have it used in jeca. By default jeca will try to guess what
to print in complex customfields, but if that's not enough, you can
write your own. Because they're custom fields, they're unique per
instance and won't apply to other people.

Adding a directory examples with a config file and a sample customfield
handler.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>