field: fixing number/float
The need for dictionary are for specific types like option, with or not
custom fields.

{ field: value } for direct fields and { field: [{subfield: value}] }
for options and more complex types, where you just need to usually pass
name or value in order to filter the right object. This can be very
confusing as, for example, DTM is a number but it's actually an option.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>